packages/tools/ussy-plan9webplumb/pyproject.toml
[project]
name = "ussy-plan9webplumb"
version = "0.1.0"
description = "Migrated from plan9webplumb"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ussy-core",
    "websockets>=11.0",
    "pyyaml>=6.0",
]

[project.scripts]
ussy-plan9webplumb = "ussy_plan9webplumb.cli:main"
plan9webplumb = "ussy_plan9webplumb.legacy:deprecated_main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ussy_plan9webplumb"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
